refactor/#155 : aop 폴더 생성
