+++ socket/src/main/java/io/eagle/aop/DistributeLock.java
@@ -1,4 +1,4 @@
-package io.eagle.common;
+package io.eagle.aop;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
